refactor(Test): 테스트코드 패키지 구조 변경 및 Lotto 접근제어자 default로 변경
