+++ src/test/java/lotto/domain/PurchaseMoneyTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.PurchaseMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
